+++ src-docs/src/views/filter_group/filter_group_example.js
@@ -1,7 +1,6 @@
 import React, { Fragment } from 'react';
 
 import {
-  Link,
 } from 'react-router';
 
 import { renderToHtml } from '../../services';
@@ -11,39 +10,114 @@ import {
 } from '../../components';
 
 import {
-  EuiCallOut,
   EuiCode,
   EuiFilterGroup,
   EuiFilterButton,
   EuiFilterSelectItem,
-  EuiSpacer,
 } from '../../../../src/components';
 
 import FilterGroup from './filter_group';
 const filterGroupSource = require('!!raw-loader!./filter_group');
 const filterGroupHtml = renderToHtml(FilterGroup);
 
+import FilterGroupSimple from './filter_group_simple';
+const filterGroupSimpleSource = require('!!raw-loader!./filter_group_simple');
+const filterGroupSimpleHtml = renderToHtml(FilterGroup);
+
+import FilterGroupMulti from './filter_group_multi';
+const filterGroupMultiSource = require('!!raw-loader!./filter_group_multi');
+const filterGroupMultiHtml = renderToHtml(FilterGroup);
+
 export const FilterGroupExample = {
   title: 'Filter Group',
-  intro: (
-    <Fragment>
-      <EuiCallOut
-        title="Demo of visual pattern only"
-        color="warning"
-      >
-        <p>
-          This documents a visual pattern used for filtering (usually page heads next to search).
-          The individual components themselves are very simple
-          and do not have much functionality on their own. If you are looking for expanded usage
-          examples please check out the <Link to="/forms/search-bar">Search Bar</Link> component which uses this more fully and
-          can give you a better example of its usage when applied to filtering.
-        </p>
-      </EuiCallOut>
-
-      <EuiSpacer size="l" />
-    </Fragment>
-  ),
   sections: [{
+    title: 'Filter buttons',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: filterGroupSimpleSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: filterGroupSimpleHtml,
+    }],
+    text: (
+      <Fragment>
+        <p>
+          Use <EuiCode>EuiFilterGroup</EuiCode> to wrap <EuiCode>EuiFilterButton</EuiCode>s
+          into a container that looks nice against form fields (like search). These buttons
+          are used in two different patterns. The most simplest use is that of an on/off pattern to show
+          whether a filter is on. Add the prop <EuiCode>withNext</EuiCode> to remove the border between it
+          and the next EuiFilterButton to visually group similar or opposite style filters.
+        </p>
+        <p>
+          Add the prop <EuiCode>withNext</EuiCode> to remove the border between it
+          and the next EuiFilterButton to visually group similar or opposite style filters.
+        </p>
+        <p>
+          Set <EuiCode>hasActiveFilters</EuiCode> to true when the filter is active.
+        </p>
+      </Fragment>
+    ),
+    props: { EuiFilterGroup, EuiFilterButton },
+    demo: <FilterGroupSimple />,
+    snippet: `<EuiFilterGroup>
+  <EuiFilterButton
+    hasActiveFilters={this.state.isFilterOn}
+    onClick={this.toggleFilter}
+  >
+    Single filter
+  </EuiFilterButton>
+</EuiFilterGroup>`
+  },
+  {
+    title: 'Multi-select',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: filterGroupMultiSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: filterGroupMultiHtml,
+    }],
+    text: (
+      <Fragment>
+        <p>
+          To provide a long list of grouped filter, use a popover for filtering an
+          array of passed items. This mostly uses standard popover mechanics, but the
+          component <EuiCode>EuiFilterSelectItem</EuiCode> is used for the items themselves.
+        </p>
+        <h3>Indicating number of filters</h3>
+        <p>
+          By passing a number to <EuiCode>numFilters</EuiCode> you can express the number of filters available.
+          When the user has applied these filter add the prop <EuiCode>hasActiveFilters</EuiCode> as before
+          and this will change the coloring of the indicator. You can also supply a number to <EuiCode>numActiveFilters</EuiCode>
+          which will change the number displayed.
+        </p>
+      </Fragment>
+    ),
+    props: { EuiFilterButton, EuiFilterSelectItem },
+    demo: <FilterGroupMulti />,
+    snippet: `<EuiFilterGroup>
+  <EuiPopover
+    button={
+      <EuiFilterButton
+        iconType="arrowDown"
+        onClick={this.onButtonClick}
+        isSelected={this.state.isPopoverOpen}
+        numFilters={items.length}
+        hasActiveFilters={true}
+        numActiveFilters={2}
+      >
+        Filters
+      </EuiFilterButton>
+    }
+    isOpen={this.state.isPopoverOpen}
+    closePopover={this.closePopover}
+  >
+    ...
+  </EuiPopover>
+</EuiFilterGroup>`
+  },
+  {
+    title: 'Layout',
     source: [{
       type: GuideSectionTypes.JS,
       code: filterGroupSource,
@@ -53,16 +127,24 @@ export const FilterGroupExample = {
     }],
     text: (
       <p>
-        Use <EuiCode>EuiFilterGroup</EuiCode> to wrap <EuiCode>EuiFilterButton</EuiCode>s
-        into a container that looks nice against form fields (like search). These buttons
-        are used in two different patterns. The first, as a simple on/off pattern to show
-        whether a setting is on. The second is as delivery for a popover for filtering an
-        array of passed items. This mostly uses standard popover mechanics, but the
-        component <EuiCode>EuiFilterSelectItem</EuiCode> is used for the items themselves.
+        By default, the bar is auto-widthed based on its contents. To expand the bar to
+        fill its parent&apos;s width add <EuiCode>fullWidth</EuiCode>. This will also set
+        each button to grow. If you do not want the button to grow, set <EuiCode>grow = false</EuiCode>.
       </p>
     ),
     components: { EuiFilterGroup },
     props: { EuiFilterGroup, EuiFilterButton, EuiFilterSelectItem },
     demo: <FilterGroup />,
+    snippet: `<EuiFilterGroup fullWidth>
+  <EuiFilterButton>
+    Single filter
+  </EuiFilterButton>
+  <EuiFilterButton grow={false} withNext>
+    On
+  </EuiFilterButton>
+  <EuiFilterButton grow={false}>
+    Off
+  </EuiFilterButton>
+</EuiFilterGroup>`,
   }],
 };